+++ branches/sayonara-832/src/GUI/library/model/LibraryItemModelAlbums.cpp
@@ -235,8 +235,8 @@ QModelIndex LibraryItemModelAlbums::getPrevRowIndexOf(QString substr, int row){
 
 
 
-QList<int> LibraryItemModelAlbums::get_discnumbers(const QModelIndex& idx){
-    if(!idx.isValid()) return QList<int>();
-    if(idx.row() < 0 || idx.row() >= _album_list.size()) return QList<int>();
+QList<quint8> LibraryItemModelAlbums::get_discnumbers(const QModelIndex& idx){
+    if(!idx.isValid()) return QList<quint8>();
+    if(idx.row() < 0 || idx.row() >= _album_list.size()) return QList<quint8>();
 	return _album_list[idx.row()].discnumbers;
 }
